(Issue) sorting is now done inside the adapter
Adapter receives issue and repository, adds the issue into an array list
(as before), but repository is inserted into a hash map. The key is the
issues url (which is unique for each issue). The adapter then sorts the
issues. When issue is clicked, the adapter sends the issue and
repository (retrieving it based in key).

Not only does moving sorting into the adapter makes more sense, but it
will allow in the future for custom sorting. In order to implement
custom sorting, appropriate method needs to be implemented in the
adapter that will sort the issues (without the need to fetch the data
again)